Added Prim's Algorithm to WG class
Got a bare bones version of Prim's Algorithm running and printing the way it should. I used the sample graph from the lecture slides on weighted graphs. If you look at the slides, my algorithm is not working correctly because it didn't use both edges from vertex A. We'll have to come back and fix it at some point.

Updated the Edge class to make it comparable for the priority queue and added a boolean variable to the Vertex class.